Launch Taipan

This launches the app

launchTaipan(questions = sampleQuestions, loadCache = FALSE,
  images = list.files("images"))

Arguments

questions
A taipan structured list object containing the questions to be asked in the app, see the global environment object sampleQuestions.

loadCache
If a cache.Rda object is found in the working directory and loadCache = TRUE, data from previous use of the app will be made avaiable. if there is no object the app will proceed from the beinnning of the image set. The default for this is FALSE.

images
The list of images located within the image folder.

Examples

# NOT RUN {
library(taipan)
tennisQuestionsSmall <- list(scene = list(

bg = list(qType = "radio",
         label = "What is the background?",
         choices = c("Crowd", "Court", "Logo wall", "Not applicable")),
shotangle = list(qType = "radio",
                label = "What angle was the image taken from?",
                choices = c("Level with players","Birds eye", "Upward angle")),
situation = list(qType = "radio",
                label = "What is the siutation occurring?",
                choices = c("Court in play", "Court player close-up","Court close-up not player",
                            "Crowd", "Off court close up of player","Transition"))),

selection = list(detect = list(qType = "radio",
                              label = "Who is the person selected?",
                              choices = c("Player","Other staff on court", "Fan")),
                glasses = list(qType = "radio",
                               label = "Is the person wearing glasses?",
                               choices = c("No", "Yes")),
                visorhat = list(qType = "radio",
                                label = "Is the person wearing a visor or hat?",
                                choices = c("No", "Yes"))))

launchTaipan(tennisQuestionsSmall)
# }